requirements.txt
Django==3.0
djangorestframework==3.11.0
django-filter==2.2.0
django-cors-headers==3.2.1